[WIP] Y axis: correct deph raycasting and correct grid displaying

+++ Editor/Shaders/grid-3.0.vert
@@ -3,7 +3,7 @@
 in vec3 position;
 
 uniform mat4 modelviewProjection;
-uniform int yPosition;
+uniform float yPosition;
 
 void main()
 {
