+++ Editor/Shaders/grid-3.3.vert
@@ -3,7 +3,7 @@
 layout(location = 0) in vec3 position;
 
 uniform mat4 modelviewProjection;
-uniform int yPosition;
+uniform float yPosition;
 
 void main()
 {
