+++ Editor/Shaders/grid.vert
@@ -1,7 +1,7 @@
 attribute vec3 position;
 
 uniform mat4 modelviewProjection;
-uniform int yPosition;
+uniform float yPosition;
 
 void main()
 {
